+++ htdocs/cabinetmed/class/cabinetmedcons.class.php
@@ -433,13 +433,10 @@ class CabinetmedCons extends CommonObject
 
 		if (! $error)
 		{
-		    if (empty($conf->global->MAIN_EXTRAFIELDS_DISABLED)) // For avoid conflicts if trigger used
-		    {
-		        $result=$this->insertExtraFields();
-		        if ($result < 0)
-		        {
-		            $error++;
-		        }
+	        $result=$this->insertExtraFields();
+	        if ($result < 0)
+	        {
+	            $error++;
 		    }
 		}