+++ htdocs/cabinetmed/core/modules/modCabinetMed.class.php
@@ -506,10 +506,14 @@ class modCabinetMed extends DolibarrModules
         'co.banque'=>'generic:Consultation'
         );   // We define here only fields that use another picto
 
+        $keyforselect='cabinetmed_cons'; $keyforaliasextra='extra'; $keyforelement='consultation';
+        include DOL_DOCUMENT_ROOT.'/core/extrafieldsinexport.inc.php';
+
         $this->export_sql_start[$r]='SELECT DISTINCT ';
         $this->export_sql_end[$r]  =' FROM '.MAIN_DB_PREFIX.'societe as s';
         $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'cabinetmed_patient as pa ON s.rowid = pa.rowid';
         $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'cabinetmed_cons as co ON s.rowid = co.fk_soc';
+        $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'cabinetmed_cons_extrafields as extra ON co.rowid = extra.fk_object';
         $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'c_typent as t ON s.fk_typent = t.id';
         $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'c_country as p ON s.fk_pays = p.rowid';
         $this->export_sql_end[$r] .=' LEFT JOIN '.MAIN_DB_PREFIX.'c_effectif as ce ON s.fk_effectif = ce.id';