+++ htdocs/cabinetmed/ChangeLog.md
@@ -4,7 +4,7 @@
 
 - Compatibility of export with PHP 7.4
 - Can set SOCIETE_DISABLE_PARENTCOMPANY to enabled use of parent company.
-
+- Add extrafields of consultation into export of patient and consultations.
 
 ## 9.0.0